style: rename login to authentication
